+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## [Added] - {PR_MERGE_DATE}
 
-- Initial release of the SCP File Transfer extension with upload and download capabilities
+- Initial release of the Rsync File Transfer extension with upload and download capabilities
 - Support for SSH config file integration
 - Support for recursive directory transfers